refactor: RecordButton test, window - Fiile, FormData 관련 목킹 setup으로 분리 및 수정
